refactor: [refactor] User Role 객체 관리용 파일 추가
1) 기존 Entity와 Auth DTO 에서 같은 코드를 사용하고 있었는데 하나의 파일로 따로 만들어 관리
