+++ src/auth/guard/roles.guard.ts
@@ -1,4 +1,4 @@
-import { Role } from '../../users/entities/user.entity';
+import { Role } from '../../users/entities/user-role';
 
 import { CanActivate, ExecutionContext, Injectable } from '@nestjs/common';
 import { Reflector } from '@nestjs/core';
