+++ src/seats/seats.controller.ts
@@ -15,7 +15,8 @@ import { SeatsService } from './seats.service';
 import { CreateSeatDto } from './dto/create-seat.dto';
 import { RolesGuard } from 'src/auth/guard/roles.guard';
 import { Roles } from 'src/utils/role.decorator';
-import { Role, User } from 'src/users/entities/user.entity';
+import { User } from 'src/users/entities/user.entity';
+import { Role } from 'src/users/entities/user-role';
 import { FileInterceptor } from '@nestjs/platform-express';
 import { Seat } from './entities/seat.entity';
 import { AuthGuard } from '@nestjs/passport';
