+++ src/users/entities/user-role.ts
@@ -1,0 +1,4 @@
+export enum Role {
+  Admin = 'Admin',
+  Customer = 'Customer',
+}
